+++ cmds/inspect.c
@@ -27,7 +27,7 @@
 #include "ioctl.h"
 #include "common/utils.h"
 #include "ctree.h"
-#include "send-utils.h"
+#include "common/send-utils.h"
 #include "disk-io.h"
 #include "cmds/commands.h"
 #include "btrfs-list.h"
