+++ libbtrfs.sym
@@ -44,7 +44,7 @@ global:
 	/* common/send-stream.h */
 	btrfs_read_and_process_send_stream;
 
-	/* send-utils.h */
+	/* common/send-utils.h */
 	subvol_uuid_search;
 	subvol_uuid_search2;
 	subvol_uuid_search_add;
